Clean up and merge tests
